+++ sakura_core/types/CType_Cpp.cpp
@@ -1472,7 +1472,7 @@ void CEditView::SmartIndent_CPP( wchar_t wcChar )
 							nBegin = n;
 							n++;
 						}else if( m_pTypeData->m_bIndentCppCommentIgnore && n + 1 < nLineLen2 && '/' == pLine2[n] && '/' == pLine2[n+1] ){
-							SCommentBlock block = { n, nLineLen2 - 1 };
+							SCommentBlock block = { static_cast<int>(n), static_cast<int>(static_cast<ssize_t>(nLineLen2) - 1) };
 							arrCommentBlock.push_back(block);
 							n = nLineLen2 - 1;
 						}
@@ -1481,7 +1481,7 @@ void CEditView::SmartIndent_CPP( wchar_t wcChar )
 						if( L'\\' == pLine2[n] ){
 							n++;
 						}else if( L'"' == pLine2[n] ){
-							SCommentBlock block = { nBegin, n };
+							SCommentBlock block = { nBegin, static_cast<int>(n) };
 							arrCommentBlock.push_back(block);
 							nBegin = -1;
 							nMode = 0;
@@ -1491,7 +1491,7 @@ void CEditView::SmartIndent_CPP( wchar_t wcChar )
 						if( L'\\' == pLine2[n] ){
 							n++;
 						}else if( L'\'' == pLine2[n] ){
-							SCommentBlock block = { nBegin, n };
+							SCommentBlock block = { nBegin, static_cast<int>(n) };
 							arrCommentBlock.push_back(block);
 							nBegin = -1;
 							nMode = 0;
@@ -1499,7 +1499,7 @@ void CEditView::SmartIndent_CPP( wchar_t wcChar )
 						break;
 					case 3:
 						if( n + 1 < nLineLen2 && '*' == pLine2[n] && '/' == pLine2[n+1] ){
-							SCommentBlock block = { nBegin, n + 1 };
+							SCommentBlock block = { nBegin, static_cast<int>(n + 1) };
 							arrCommentBlock.push_back(block);
 							nBegin = -1;
 							nMode = 0;
@@ -1512,7 +1512,7 @@ void CEditView::SmartIndent_CPP( wchar_t wcChar )
 				}
 				if( 0 < nBegin ){
 					// 終わりのない文字列orコメント
-					SCommentBlock block = { nBegin, nLineLen2 };
+					SCommentBlock block = { nBegin, static_cast<int>(nLineLen2) };
 					arrCommentBlock.push_back(block);
 				}
 				if( j == GetCaret().GetCaretLogicPos().GetY2() ){
